new scenario in reports
